Removed script.js Fixed dropdown styling (removed padding, centered text) Repaired broken chart data (parseInt --> parseFloat)

+++ myChart.js
@@ -135,7 +135,7 @@ function getHistorical(queryURL) {
     var delta = data.length - arr.length
     var i = delta
     for (i; i < data.length; i++) {
-      var price = parseInt(data[i].priceUsd).toFixed(3);
+      var price = parseFloat(data[i].priceUsd).toFixed(3);
       arr[i-delta-1] = price
       labels[i-delta-1] = globalCoin
     }
